requirements.txt
# ── Core ──────────────────────────────────────────────────
torch>=2.4.0
torchvision>=0.19.0
transformers>=5.5.0
accelerate>=1.0.0
pillow>=10.0.0
av>=12.0.0
huggingface-hub>=0.20.0
numpy

# ── Video output ──────────────────────────────────────────
moviepy>=2.0.0

# ── Fine-tuning (QLoRA) ──────────────────────────────────
peft>=0.13.0
bitsandbytes>=0.44.0
trl>=0.12.0
datasets>=3.0.0
pyyaml>=6.0

# ── Desktop GUI (optional -- install for tools/gui.py) ───
# PySide6>=6.8.0
# matplotlib>=3.9.0
# windows-capture>=2.0.0       # Windows only

# ── System dependency (not pip) ───────────────────────────
# sudo apt-get install -y fonts-noto-cjk

# ── Optional: audio transcription (--audio flag) ─────────
# faster-whisper>=1.0.0
# openai-whisper>=20231117
